backend/requirements.txt
fastapi==0.128.0
uvicorn==0.30.6
gunicorn==22.0.0
python-dotenv==1.0.1
alembic==1.13.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.3
requests==2.32.4
httpx==0.27.2
aiofiles==24.1.0
python-multipart==0.0.22
openpyxl==3.1.5
reportlab==4.2.5
PyMuPDF==1.24.9
redis==5.0.8
celery==5.4.0
chromadb==0.5.5
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
openai>=1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
rarfile==4.1
boto3==1.35.28
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-jose==3.4.0
pydantic-settings==2.3.4
networkx==3.3
statsmodels==0.14.2
cryptography==44.0.1
